refactor core #171
