+++ src/lib/core/dto/did-dto.ts
