+++ src/lib/core/exceptions/auth-exceptions.ts
