+++ src/lib/core/port/primary/set-x509-login-session-input-port.ts
@@ -1,4 +1,4 @@
-import { SetX509LoginSessionRequest } from "../../data/usecase-models/set-x509-login-session-usecase-models";
+import { SetX509LoginSessionRequest } from "../../usecase-models/set-x509-login-session-usecase-models";
 
 /**
  * @interface SetX509LoginSessionInputPort to set trigger setting up the IronSession in the Presenter
