+++ src/lib/core/port/primary/userpass-login-input-port.ts
@@ -1,4 +1,4 @@
-import { UserpassLoginRequest } from "../../data/usecase-models/userpass-login-usecase-models";
+import { UserpassLoginRequest } from "../../usecase-models/userpass-login-usecase-models";
 
 /**
  * InputPort for UserPassLogin workflow. This is implemented by {@link UserPassLoginUseCase}
