+++ src/lib/core/port/secondary/account-gateway-output-port.ts
@@ -1,4 +1,4 @@
-import { AccountAttributesDTO } from "../../data/dto/account-dto";
+import { AccountAttributesDTO } from "../../dto/account-dto";
 
 export default interface AccountGatewayOutputPort {
     listAccountAttributes(account: string, rucioAuthToken: string): Promise<AccountAttributesDTO>
