+++ src/lib/infrastructure/data/view-model/list-did.d.ts
@@ -1,5 +1,5 @@
 import { BaseViewModel } from "@/lib/common/base-components/view-models";
-import { ListDIDsResponse } from "@/lib/core/use-case/list-dids/list-dids-usecase-models";
+import { ListDIDsResponse } from "@/lib/core/usecase-models/list-dids-usecase-models";
 import { DID } from "@/lib/core/entity/rucio";
 
 export interface ListDIDsViewModel extends DID, BaseViewModel {
