+++ src/lib/infrastructure/gateway/did-gateway/did-gateway.ts
@@ -1,5 +1,5 @@
 import { HTTPRequest } from '@/lib/common/http'
-import { DIDDTO, ListDIDDTO } from '@/lib/core/data/dto/did-dto'
+import { DIDDTO, ListDIDDTO } from '@/lib/core/dto/did-dto'
 import { DIDType } from '@/lib/core/entity/rucio'
 import DIDGatewayOutputPort from '@/lib/core/port/secondary/did-gateway-output-port'
 import { injectable } from 'inversify'
