+++ src/lib/infrastructure/gateway/did-gateway/endpoints/get-did-endpoint.ts
@@ -1,6 +1,6 @@
 import { BaseEndpoint } from "@/lib/common/base-components/gateway-endpoints"
 import { HTTPRequest } from "@/lib/common/http";
-import { DIDDTO } from "@/lib/core/data/dto/did-dto";
+import { DIDDTO } from "@/lib/core/dto/did-dto";
 import { DIDType } from "@/lib/core/entity/rucio";
 import { Response } from "node-fetch";
